% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiddenSafetynet2025-package.R
\docType{package}
\name{HiddenSafetynet2025-package}
\alias{HiddenSafetynet2025}
\alias{HiddenSafetynet2025-package}
\title{HiddenSafetynet2025: HiddenSafetynet2025}
\description{
Replication Package for Hidden Safety Net of Underutilized Supplemental Insurance in US Agriculture.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/you/HiddenSafetynet2025}
  \item Report bugs at \url{https://github.com/you/HiddenSafetynet2025/issues}
}

}
\keyword{internal}
